feat / plugin permissions
