+++ Cargo.toml
@@ -11,23 +11,7 @@ incremental = false
 codegen-units = 1
 incremental = false
 
-[profile.release.package.vectis-govec]
-codegen-units = 1
-incremental = false
-
 [profile.release.package.vectis-plugin-registry]
-codegen-units = 1
-incremental = false
-
-[profile.release.package.vectis-dao-tunnel]
-codegen-units = 1
-incremental = false
-
-[profile.release.package.vectis-remote-tunnel]
-codegen-units = 1
-incremental = false
-
-[profile.release.package.vectis-remote-factory]
 codegen-units = 1
 incremental = false
