+++ cli/deploy/juno_testnet-deployInfo.json
@@ -1,8 +1,8 @@
 {
-  "PluginCommittee": "juno15m2y8z8eq4qe6vllexzs30u4wskxlava6uascnhdxcm880x7c6kq0jg0d5",
-  "PluginCommitteeGroup": "juno1k245anxsl84ez4hwnq79ytgcc4g980ghnlqzd7daf5ecnsr4gegskezj0d",
-  "VectisCommittee": "juno1jjja02r3l0yuxhyml6wz4r3qyzn3e5vc8gsrp8tcpcja4xscpahsa039ys",
-  "VectisCommitteeGroup": "juno1v87e5zu6dkwmuk7umguecwklx6xjucze8tqfmaqhsxaxl8kjqw8q6kt8un",
-  "PluginRegistry": "juno1gmgfhzc9a97trnk5l7wastq9m5lz5n6zteud8uefndryfrrek74s73t9xn",
-  "Factory": "juno1lptxt8xv0glt7r5zslngu6gu9tu5trkkc7g630e0vlfk00ccnhqq0x5txt"
+  "PluginCommittee": "juno1ptq2g9qtkm59gpu0dvds63qrn68qhevklzj2rfm2352vndtvepfq9kh4ff",
+  "PluginCommitteeGroup": "juno1ghvy3y7ms9ygyzmcc3rrzjk40pduf6pu8jefhvg08ml2zrtm3tqqpw0zgq",
+  "VectisCommittee": "juno1cw3225vrncnmyvfjw4lqyknynmcgctpra3f6mmce6lsfn5trra4sp88w42",
+  "VectisCommitteeGroup": "juno1n6hgtkdm0grj8z0847tqmwc47m2kjf2uwejs9rl579mqjvjyu98sjlfj9q",
+  "PluginRegistry": "juno184k0kdmwefxdr7zypf4y0w5967v3kum3kf607lpg820nsm6d8w6qymad27",
+  "Factory": "juno19s4se2k97dpxg9m3auazknfvh72hgqwws8uuauenkawu65lqxx4q2kkhvh"
 }
