+++ cli/utils/fees.ts
@@ -2,8 +2,9 @@ import { coin, calculateFee, GasPrice } from "@cosmjs/stargate";
 import { Coin } from "../interfaces/Factory.types";
 import type { Chain } from "../config/chains";
 
-export const getDefaultGasPrice = (chain: Chain) => GasPrice.fromString(chain.gasPrice + chain.feeToken);
+export const walletInitialFunds = (chain: Chain) => coin(1500000, chain.feeToken) as Coin;
 
+export const getDefaultGasPrice = (chain: Chain) => GasPrice.fromString(chain.gasPrice + chain.feeToken);
 export const getInitialFactoryBalance = (chain: Chain) => coin(1_000_000, chain.feeToken) as Coin;
 
 export const getDefaultUploadFee = (chain: Chain) => calculateFee(55_500_000, getDefaultGasPrice(chain));
@@ -16,6 +17,6 @@ export const getDefaultMigrateFee = (chain: Chain) => calculateFee(1_200_000, ge
 export const getDefaultUpdateAdminFee = (chain: Chain) => calculateFee(800_000, getDefaultGasPrice(chain));
 export const getDefaultClearAdminFee = (chain: Chain) => calculateFee(800_000, getDefaultGasPrice(chain));
 
-export const walletInitialFunds = (chain: Chain) => coin(0, chain.feeToken) as Coin;
+export const walletCreationFee = (chain: Chain) => coin(0, chain.feeToken) as Coin;
 export const pluginRegInstallFee = (chain: Chain) => coin(0, chain.feeToken) as Coin;
 export const pluginRegRegistryFee = (chain: Chain) => coin(0, chain.feeToken) as Coin;
