+++ contracts/factory/schema/vectis-factory.json
@@ -183,6 +183,62 @@
           }
         },
         "additionalProperties": false
+      },
+      {
+        "type": "object",
+        "required": [
+          "update_controller"
+        ],
+        "properties": {
+          "update_controller": {
+            "type": "object",
+            "required": [
+              "new_controller",
+              "old_controller"
+            ],
+            "properties": {
+              "new_controller": {
+                "$ref": "#/definitions/Addr"
+              },
+              "old_controller": {
+                "$ref": "#/definitions/Addr"
+              }
+            },
+            "additionalProperties": false
+          }
+        },
+        "additionalProperties": false
+      },
+      {
+        "type": "object",
+        "required": [
+          "update_guardians"
+        ],
+        "properties": {
+          "update_guardians": {
+            "type": "object",
+            "required": [
+              "new_guardians",
+              "old_guardians"
+            ],
+            "properties": {
+              "new_guardians": {
+                "type": "array",
+                "items": {
+                  "$ref": "#/definitions/Addr"
+                }
+              },
+              "old_guardians": {
+                "type": "array",
+                "items": {
+                  "$ref": "#/definitions/Addr"
+                }
+              }
+            },
+            "additionalProperties": false
+          }
+        },
+        "additionalProperties": false
       }
     ],
     "definitions": {
@@ -275,7 +331,7 @@
             "description": "A List of keys can act as guardian for",
             "type": "array",
             "items": {
-              "type": "string"
+              "$ref": "#/definitions/Addr"
             }
           },
           "guardians_multisig": {
@@ -295,16 +351,9 @@
       "MultiSig": {
         "type": "object",
         "required": [
-          "multisig_initial_funds",
           "threshold_absolute_count"
         ],
         "properties": {
-          "multisig_initial_funds": {
-            "type": "array",
-            "items": {
-              "$ref": "#/definitions/Coin"
-            }
-          },
           "threshold_absolute_count": {
             "description": "Only Fixed multisig is supported in this version",
             "type": "integer",
@@ -486,9 +535,57 @@
           }
         },
         "additionalProperties": false
+      },
+      {
+        "description": "Returns the wallet controlled by this controller",
+        "type": "object",
+        "required": [
+          "controller_wallets"
+        ],
+        "properties": {
+          "controller_wallets": {
+            "type": "object",
+            "required": [
+              "controller"
+            ],
+            "properties": {
+              "controller": {
+                "$ref": "#/definitions/Addr"
+              }
+            },
+            "additionalProperties": false
+          }
+        },
+        "additionalProperties": false
+      },
+      {
+        "description": "Returns the wallet with this guardian",
+        "type": "object",
+        "required": [
+          "wallets_with_guardian"
+        ],
+        "properties": {
+          "wallets_with_guardian": {
+            "type": "object",
+            "required": [
+              "guardian"
+            ],
+            "properties": {
+              "guardian": {
+                "$ref": "#/definitions/Addr"
+              }
+            },
+            "additionalProperties": false
+          }
+        },
+        "additionalProperties": false
       }
     ],
     "definitions": {
+      "Addr": {
+        "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
+        "type": "string"
+      },
       "CodeIdType": {
         "type": "string",
         "enum": [
@@ -507,6 +604,20 @@
       "type": "integer",
       "format": "uint64",
       "minimum": 0.0
+    },
+    "controller_wallets": {
+      "$schema": "http://json-schema.org/draft-07/schema#",
+      "title": "Array_of_Addr",
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/Addr"
+      },
+      "definitions": {
+        "Addr": {
+          "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
+          "type": "string"
+        }
+      }
     },
     "deployer_addr": {
       "$schema": "http://json-schema.org/draft-07/schema#",
@@ -555,6 +666,20 @@
       "type": "integer",
       "format": "uint64",
       "minimum": 0.0
+    },
+    "wallets_with_guardian": {
+      "$schema": "http://json-schema.org/draft-07/schema#",
+      "title": "Array_of_Addr",
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/Addr"
+      },
+      "definitions": {
+        "Addr": {
+          "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
+          "type": "string"
+        }
+      }
     }
   }
 }
