+++ contracts/proxy/schema/vectis-proxy.json
@@ -30,6 +30,10 @@
     },
     "additionalProperties": false,
     "definitions": {
+      "Addr": {
+        "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
+        "type": "string"
+      },
       "Coin": {
         "type": "object",
         "required": [
@@ -90,7 +94,7 @@
             "description": "A List of keys can act as guardian for",
             "type": "array",
             "items": {
-              "type": "string"
+              "$ref": "#/definitions/Addr"
             }
           },
           "guardians_multisig": {
@@ -110,16 +114,9 @@
       "MultiSig": {
         "type": "object",
         "required": [
-          "multisig_initial_funds",
           "threshold_absolute_count"
         ],
         "properties": {
-          "multisig_initial_funds": {
-            "type": "array",
-            "items": {
-              "$ref": "#/definitions/Coin"
-            }
-          },
           "threshold_absolute_count": {
             "description": "Only Fixed multisig is supported in this version",
             "type": "integer",
@@ -722,7 +719,7 @@
             "description": "A List of keys can act as guardian for",
             "type": "array",
             "items": {
-              "type": "string"
+              "$ref": "#/definitions/Addr"
             }
           },
           "guardians_multisig": {
@@ -916,16 +913,9 @@
       "MultiSig": {
         "type": "object",
         "required": [
-          "multisig_initial_funds",
           "threshold_absolute_count"
         ],
         "properties": {
-          "multisig_initial_funds": {
-            "type": "array",
-            "items": {
-              "$ref": "#/definitions/Coin"
-            }
-          },
           "threshold_absolute_count": {
             "description": "Only Fixed multisig is supported in this version",
             "type": "integer",
@@ -1422,20 +1412,9 @@
         }
       ],
       "definitions": {
-        "Coin": {
-          "type": "object",
-          "required": [
-            "amount",
-            "denom"
-          ],
-          "properties": {
-            "amount": {
-              "$ref": "#/definitions/Uint128"
-            },
-            "denom": {
-              "type": "string"
-            }
-          }
+        "Addr": {
+          "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
+          "type": "string"
         },
         "Expiration": {
           "description": "Expiration represents a point in time when some event happens. It can compare with a BlockInfo and will return is_expired() == true once the condition is hit (and for every block in the future)",
@@ -1494,7 +1473,7 @@
               "description": "A List of keys can act as guardian for",
               "type": "array",
               "items": {
-                "type": "string"
+                "$ref": "#/definitions/Addr"
               }
             },
             "guardians_multisig": {
@@ -1515,13 +1494,14 @@
           "type": "object",
           "required": [
             "activate_at",
-            "guardians"
+            "new_guardians",
+            "old_guardians"
           ],
           "properties": {
             "activate_at": {
               "$ref": "#/definitions/Expiration"
             },
-            "guardians": {
+            "new_guardians": {
               "$ref": "#/definitions/Guardians"
             },
             "new_multisig_code_id": {
@@ -1531,6 +1511,12 @@
               ],
               "format": "uint64",
               "minimum": 0.0
+            },
+            "old_guardians": {
+              "type": "array",
+              "items": {
+                "$ref": "#/definitions/Addr"
+              }
             }
           },
           "additionalProperties": false
@@ -1538,16 +1524,9 @@
         "MultiSig": {
           "type": "object",
           "required": [
-            "multisig_initial_funds",
             "threshold_absolute_count"
           ],
           "properties": {
-            "multisig_initial_funds": {
-              "type": "array",
-              "items": {
-                "$ref": "#/definitions/Coin"
-              }
-            },
             "threshold_absolute_count": {
               "description": "Only Fixed multisig is supported in this version",
               "type": "integer",
@@ -1564,10 +1543,6 @@
               "$ref": "#/definitions/Uint64"
             }
           ]
-        },
-        "Uint128": {
-          "description": "A thin wrapper around u128 that is using strings for JSON encoding/decoding, such that the full u128 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.\n\n# Examples\n\nUse `from` to create instances of this and `u128` to get the value out:\n\n``` # use cosmwasm_std::Uint128; let a = Uint128::from(123u128); assert_eq!(a.u128(), 123);\n\nlet b = Uint128::from(42u64); assert_eq!(b.u128(), 42);\n\nlet c = Uint128::from(70u32); assert_eq!(c.u128(), 70); ```",
-          "type": "string"
         },
         "Uint64": {
           "description": "A thin wrapper around u64 that is using strings for JSON encoding/decoding, such that the full u64 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.\n\n# Examples\n\nUse `from` to create instances of this and `u64` to get the value out:\n\n``` # use cosmwasm_std::Uint64; let a = Uint64::from(42u64); assert_eq!(a.u64(), 42);\n\nlet b = Uint64::from(70u32); assert_eq!(b.u64(), 70); ```",
